# frontend/landings/agivys/landing/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AgiVysSystem API",
    "description": "API para gestão do ecossistema AgiVysSystem",
    "version": "v1"
  },
  "servers": [
    {
      "url": "https://joederblanca.com.br/agivys-api"
    }
  ],
  "paths": {
...
    "/api/v1/companies/create-with-address": {
      "post": {
        "tags": [
          "Company"
        ],
        "summary": "Cadastro Completo de Empresa (com Endereço)",
        "description": "Cadastra simultaneamente uma nova empresa (Company) e seu respectivo endereço (CompanyAddress) dentro de uma única transação no banco de dados.\r\nEm caso de falha em qualquer etapa, a operação inteira é desfeita (Rollback).",
        "requestBody": {
          "description": "Dados consolidados da empresa e do endereço.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCompanyWithAddressDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCompanyWithAddressDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCompanyWithAddressDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Empresa e endereço cadastrados com sucesso."
          },
          "400": {
            "description": "Dados inválidos ou CNPJ já existente."
          },
          "401": {
            "description": "Usuário não autenticado."
          },
          "500": {
            "description": "Erro interno do servidor ao salvar os dados."
          }
        }
      }
    },
    "/api/v1/companies": {
      "post": {
        "tags": [
          "Company"
        ],
        "summary": "Cadastro Simplificado de Empresa",
        "description": "Cria uma nova empresa vinculada ao usuário proprietário especificado. \r\nO CNPJ é validado para garantir que não haja duplicidade na plataforma.",
        "requestBody": {
          "description": "Dados básicos da empresa (Nome, CNPJ, Logo).",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCompanyDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCompanyDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCompanyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Empresa cadastrada com sucesso."
          },
          "400": {
            "description": "Dados inválidos ou CNPJ já cadastrado."
          },
          "401": {
            "description": "Usuário não autenticado."
          },
          "500": {
            "description": "Erro interno do servidor."
          }
        }
      }
    },
    "/api/v1/companies/owner/{userId}": {
      "get": {
        "tags": [
          "Company"
        ],
        "summary": "Listagem de Empresas por Proprietário",
        "description": "Recupera todas as empresas cujo `UserOwnerId` corresponde ao ID fornecido.\r\nÉ comumente utilizado para preencher o seletor de contextos do Dashboard.",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID numérico do usuário proprietário.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de empresas recuperada com sucesso."
          },
          "401": {
            "description": "Usuário não autenticado."
          },
          "500": {
            "description": "Erro interno do servidor."
          }
        }
      }
    },
    "/api/v1/companies/{id}": {
      "put": {
        "tags": [
          "Company"
        ],
        "summary": "Atualização de Empresa",
        "description": "Atualiza dados cadastrais básicos de uma empresa.\r\n**Regra de Segurança**: Somente o próprio dono (UserOwnerId) da empresa possui permissão para atualizá-la (validação feita cruzando o ID do BD com o do Token).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID numérico único da empresa.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "Dados atualizados da empresa.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCompanyDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCompanyDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCompanyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Empresa atualizada com sucesso."
          },
          "400": {
            "description": "Payload inválido."
          },
          "401": {
            "description": "Usuário não autenticado."
          },
          "403": {
            "description": "Acesso negado. O usuário não é o dono da empresa."
          },
          "404": {
            "description": "Empresa não localizada."
          },
          "500": {
            "description": "Erro interno do servidor ao atualizar a empresa."
          }
        }
      },
      "delete": {
        "tags": [
          "Company"
        ],
        "summary": "Exclusão de Empresa",
        "description": "Remove completamente o registro da empresa. Esta é uma deleção física no banco de dados.\r\nÉ altamente recomendável garantir que a empresa não possua registros financeiros dependentes antes da deleção.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID numérico único da empresa.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Empresa removida com sucesso."
          },
          "401": {
            "description": "Usuário não autenticado."
          },
          "404": {
            "description": "Empresa não localizada."
          },
          "500": {
            "description": "Erro interno do servidor ao tentar deletar a empresa (ex: violação de chave estrangeira)."
          }
        }
      }
    },
    "/api/v1/companies/{companyId}/addresses": {
      "post": {
        "tags": [
          "Company"
        ],
        "summary": "Cadastro de Endereço Empresarial",
        "description": "Vincula um novo endereço a uma empresa existente. Uma empresa pode ter múltiplos endereços (ex: Matriz, Filial, Galpão).\r\n**Regra de Segurança**: Apenas o dono da empresa pode adicionar endereços à mesma.",
        "parameters": [
          {
            "name": "companyId",
            "in": "path",
            "description": "ID único numérico da empresa associada.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "Dados do endereço.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompanyAddressDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CompanyAddressDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CompanyAddressDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Endereço cadastrado e vinculado com sucesso à empresa."
          },
          "400": {
            "description": "Dados de endereço inválidos ou malformados."
          },
          "401": {
            "description": "Usuário não autenticado."
          },
          "403": {
            "description": "Permissão negada. O usuário não é o dono da empresa."
          },
          "404": {
            "description": "Empresa não encontrada para realizar o vínculo."
          },
          "500": {
            "description": "Erro interno do servidor."
          }
        }
      },
      "get": {
        "tags": [
          "Company"
        ],
        "summary": "Listagem de Endereços Empresariais",
        "description": "Recupera a lista completa de endereços atrelados a uma empresa específica.\r\nSomente o dono da empresa pode visualizar seus respectivos endereços.",
        "parameters": [
          {
            "name": "companyId",
            "in": "path",
            "description": "ID único numérico da empresa.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de endereços recuperada com sucesso."
          },
          "401": {
            "description": "Usuário não autenticado."
          },
          "403": {
            "description": "Permissão negada. O usuário não é o dono da empresa."
          },
          "404": {
            "description": "A empresa não foi localizada."
          },
          "500": {
            "description": "Erro interno do servidor."
          }
        }
      }
    },
    "/api/v1/companies/addresses/{addressId}": {
      "put": {
        "tags": [
          "Company"
        ],
        "summary": "Atualização de Endereço Empresarial",
        "description": "Edita as informações de um endereço de empresa já existente.\r\n**Regra de Segurança**: Esta operação valida cruzadamente se o usuário logado possui a empresa dona do endereço.",
        "parameters": [
          {
            "name": "addressId",
            "in": "path",
            "description": "ID numérico único do endereço a ser modificado.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "Novos dados completos do endereço.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompanyAddressDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CompanyAddressDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CompanyAddressDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Endereço atualizado com sucesso."
          },
          "400": {
            "description": "Dados inválidos enviados no payload."
          },
          "401": {
            "description": "Usuário não autenticado."
          },
          "403": {
            "description": "Permissão negada. Endereço pertence a uma empresa de outro usuário."
          },
          "404": {
            "description": "Endereço não localizado."
          },
          "500": {
            "description": "Erro interno do servidor."
          }
        }
      },
      "delete": {
        "tags": [
          "Company"
        ],
        "summary": "Exclusão de Endereço Empresarial",
        "description": "Remove definitivamente um endereço vinculado a uma empresa.\r\nValida previamente se a pessoa que está executando a exclusão é o proprietário da empresa detentora do endereço.",
        "parameters": [
          {
            "name": "addressId",
            "in": "path",
            "description": "ID numérico único do endereço a ser removido.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Endereço removido permanentemente."
          },
          "401": {
            "description": "Usuário não autenticado."
          },
          "403": {
            "description": "Permissão negada. O usuário não tem acesso a este endereço."
          },
          "404": {
            "description": "Endereço não localizado."
          },
          "500": {
            "description": "Erro interno do servidor ao tentar excluir o endereço."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CompanyAddressDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "nullable": true
          },
          "zipCode": {
            "type": "string",
            "nullable": true
          },
          "street": {
            "type": "string",
            "nullable": true
          },
          "number": {
            "type": "string",
            "nullable": true
          },
          "complement": {
            "type": "string",
            "nullable": true
          },
          "neighborhood": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "state": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CreateCompanyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "cnpj": {
            "type": "string",
            "nullable": true
          },
          "logoUrl": {
            "type": "string",
            "nullable": true
          },
          "userOwnerId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "CreateCompanyWithAddressDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "cnpj": {
            "type": "string",
            "nullable": true
          },
          "logoUrl": {
            "type": "string",
            "nullable": true
          },
          "userOwnerId": {
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "zipCode": {
            "type": "string",
            "nullable": true
          },
          "street": {
            "type": "string",
            "nullable": true
          },
          "number": {
            "type": "string",
            "nullable": true
          },
          "complement": {
            "type": "string",
            "nullable": true
          },
          "neighborhood": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "state": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UpdateCompanyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "logoUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Cole APENAS o token JWT.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}